requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
torch>=2.2.0
transformers>=4.40.0
xgboost>=2.0
shap>=0.44
mne>=1.6
antropy>=0.1.6
groq>=0.9.0
plotly>=5.18
seaborn>=0.13
matplotlib>=3.8
pywavelets>=1.6
imbalanced-learn>=0.12
optuna>=3.5
wandb
pyyaml>=6.0
